make testにおいてIL出力と想定ケースが同じ場合、コンパイルと実行をスキップするよう変更

+++ test/FrontEndTest.cs
@@ -44,7 +44,7 @@ namespace Roku.Tests
 
         public (bool Found, string Text) GetLineContent(string[] lines, string start_line, string end_line) => GetLineContent(lines, x => x.StartsWith(start_line), x => x.StartsWith(end_line));
 
-        [Test]
+        [Test, Order(1)]
         public void CompileTest()
         {
             var failed = new List<(string Path, string Message)>(
@@ -53,7 +53,11 @@ namespace Roku.Tests
                     var filename = Path.GetFileName(src);
                     var txt = File.ReadAllText(src);
                     var lines = txt.SplitLine();
-                    var il = Path.Combine(ObjDir, Path.GetFileNameWithoutExtension(src) + ".il");
+                    var testname = Path.GetFileNameWithoutExtension(src);
+                    var il = Path.Combine(ObjDir, testname + ".il");
+                    var skip = Path.Combine(ObjDir, testname + ".testskip");
+
+                    File.Delete(skip);
                     try
                     {
                         FrontEnd.Compile(new StringReader(txt), il, new string[] { "System.Runtime" });
@@ -66,6 +70,8 @@ namespace Roku.Tests
                         var il_src = File.ReadAllText(il).Trim();
 
                         if (valid.Text.Trim() != il_src) return Tuple.Create(filename, "il make a difference");
+
+                        File.WriteAllText(skip, "true");
                     }
                     catch (Exception ex)
                     {
@@ -91,7 +97,7 @@ namespace Roku.Tests
             if (failed.Count > 0) Assert.Fail(failed.Map(x => $"{x.Path}: {x.Message}").Join("\n"));
         }
 
-        [Test]
+        [Test, Order(2)]
         public void MakeTestCaseTest()
         {
             System.Text.Encoding.RegisterProvider(System.Text.CodePagesEncodingProvider.Instance);
@@ -105,15 +111,7 @@ namespace Roku.Tests
                 var err_ = Path.Combine(ObjDir, testname + ".testerr");
                 var args_ = Path.Combine(ObjDir, testname + ".testargs");
                 var name_ = Path.Combine(ObjDir, testname + ".testname");
-
-                var txt = File.ReadAllText(src);
-                var lines = txt.SplitLine().Map(x => x.TrimStart()).ToArray();
-
-                var in_p = lines.Where(x => x.StartsWith("#<=")).Map(x => x[3..] + "\r\n").Join();
-                var out_p = lines.Where(x => x.StartsWith("#=>")).Map(x => x[3..] + "\r\n").Join();
-                var err_p = lines.Where(x => x.StartsWith("#=2>")).Map(x => x[4..] + "\r\n").Join();
-                var args_p = lines.Where(x => x.StartsWith("##*")).Map(x => x[3..]).Join(" ");
-                var name_p = GetLineContent(lines, x => x == "###", x => x == "###");
+                var skip = Path.Combine(ObjDir, testname + ".testskip");
 
                 File.Delete(in_);
                 File.Delete(out_);
@@ -121,10 +119,23 @@ namespace Roku.Tests
                 File.Delete(args_);
                 File.Delete(name_);
 
-                File.WriteAllText(in_, in_p);
-                File.WriteAllText(out_, out_p);
-                File.WriteAllText(err_, err_p);
-                File.WriteAllText(args_, args_p);
+                var txt = File.ReadAllText(src);
+                var lines = txt.SplitLine().Map(x => x.TrimStart()).ToArray();
+
+                if (!File.Exists(skip))
+                {
+                    var in_p = lines.Where(x => x.StartsWith("#<=")).Map(x => x[3..] + "\r\n").Join();
+                    var out_p = lines.Where(x => x.StartsWith("#=>")).Map(x => x[3..] + "\r\n").Join();
+                    var err_p = lines.Where(x => x.StartsWith("#=2>")).Map(x => x[4..] + "\r\n").Join();
+                    var args_p = lines.Where(x => x.StartsWith("##*")).Map(x => x[3..]).Join(" ");
+
+                    File.WriteAllText(in_, in_p);
+                    File.WriteAllText(out_, out_p);
+                    File.WriteAllText(err_, err_p);
+                    File.WriteAllText(args_, args_p);
+                }
+
+                var name_p = GetLineContent(lines, x => x == "###", x => x == "###");
                 File.WriteAllText(name_, Path.GetFileNameWithoutExtension(src) + (name_p.Found ? $" {SubstringAsByte(name_p.Text.SplitLine().Take(2).Join(" "), 79 - testname.Length, sjis)}" : ""));
             });
             Assert.Pass();
